rename Adapter from liftoff to vungle
