+++ adapters/vungle/vungle_test.go
@@ -1,4 +1,4 @@
-package liftoff
+package vungle
 
 import (
 	"testing"
@@ -10,12 +10,12 @@ import (
 
 func TestJsonSamples(t *testing.T) {
 	conf := config.Adapter{
-		Endpoint: "https://liftoff.io/bit/t",
+		Endpoint: "https://vungle.io/bit/t",
 	}
-	bidder, buildErr := Builder(openrtb_ext.BidderLiftoff, conf, config.Server{ExternalUrl: "http://hosturl.com", GvlID: 667, DataCenter: "2"})
+	bidder, buildErr := Builder(openrtb_ext.BidderVungle, conf, config.Server{ExternalUrl: "http://hosturl.com", GvlID: 667, DataCenter: "2"})
 	if buildErr != nil {
 		t.Fatalf("Builder returned unexpected error %v", buildErr)
 	}
 
-	adapterstest.RunJSONBidderTest(t, "liftofftest", bidder)
+	adapterstest.RunJSONBidderTest(t, "vungletest", bidder)
 }
