+++ adapters/vungle/vungletest/exemplary/app_video_rewarded.json
@@ -42,7 +42,7 @@
   "httpCalls": [
     {
       "expectedRequest": {
-        "uri": "https://liftoff.io/bit/t",
+        "uri": "https://vungle.io/bit/t",
         "headers": {
           "Content-Type": ["application/json"],
           "Accept": ["application/json"],
