+++ adapters/vungle/vungletest/exemplary/site_video_instl.json
@@ -1,8 +1,8 @@
 {
   "mockBidRequest": {
     "id": "test-request-id",
-    "app": {
-      "bundle": "com.prebid"
+    "site": {
+      "name": "com.prebid"
     },
     "device": {
       "ifa": "87857b31-8942-4646-ae80-ab9c95bf3fab"
@@ -45,7 +45,7 @@
   "httpCalls": [
     {
       "expectedRequest": {
-        "uri": "https://liftoff.io/bit/t",
+        "uri": "https://vungle.io/bit/t",
         "headers": {
           "Content-Type": ["application/json"],
           "Accept": ["application/json"],
@@ -54,8 +54,7 @@
         "body": {
           "id": "test-request-id",
           "app": {
-            "id": "123",
-            "bundle": "com.prebid"
+            "id": "123"
           },
           "device": {
             "ifa": "87857b31-8942-4646-ae80-ab9c95bf3fab"
