+++ adapters/vungle/vungletest/exemplary/site_video_rewarded.json
@@ -1,8 +1,8 @@
 {
   "mockBidRequest": {
     "id": "test-request-id",
-    "app": {
-      "bundle": "com.prebid"
+    "site": {
+      "name": "com.prebid"
     },
     "device": {
       "ifa": "87857b31-8942-4646-ae80-ab9c95bf3fab"
@@ -25,6 +25,9 @@
           }
         },
         "ext": {
+          "prebid": {
+            "is_rewarded_inventory": 1
+          },
           "bidder": {
             "app_store_id": "123",
             "placement_reference_id": "123"
@@ -39,7 +42,7 @@
   "httpCalls": [
     {
       "expectedRequest": {
-        "uri": "https://liftoff.io/bit/t",
+        "uri": "https://vungle.io/bit/t",
         "headers": {
           "Content-Type": ["application/json"],
           "Accept": ["application/json"],
@@ -48,8 +51,7 @@
         "body": {
           "id": "test-request-id",
           "app": {
-            "id": "123",
-            "bundle": "com.prebid"
+            "id": "123"
           },
           "device": {
             "ifa": "87857b31-8942-4646-ae80-ab9c95bf3fab"
@@ -73,7 +75,9 @@
                 }
               },
               "ext": {
-                "prebid": null,
+                "prebid": {
+                  "is_rewarded_inventory": 1
+                },
                 "bidder": {
                   "app_store_id": "123",
                   "placement_reference_id": "123"
@@ -120,18 +124,18 @@
   ],
   "expectedBidResponses": [
     {
-      "currency":"USD",
-      "bids":[
+      "currency": "USD",
+      "bids": [
         {
           "bid": {
             "id": "1",
             "impid": "test-imp-id",
             "adid": "11110126",
-            "price": 0.500000,
+            "price": 0.5,
             "adm": "some-test-ad",
             "crid": "test-crid",
-            "h": 250,
-            "w": 300
+            "w": 300,
+            "h": 250
           },
           "type": "video",
           "seat": "seat-id"
