+++ adapters/vungle/vungletest/supplemental/no_site_or_app_video_rewarded.json
