+++ adapters/vungle/vungletest/supplemental/response_code_204.json
@@ -1,8 +1,8 @@
 {
   "mockBidRequest": {
     "id": "test-request-id",
-    "site": {
-      "name": "com.prebid"
+    "app": {
+      "bundle": "com.prebid"
     },
     "device": {
       "ifa": "87857b31-8942-4646-ae80-ab9c95bf3fab"
@@ -24,7 +24,6 @@
             "foo": "bar"
           }
         },
-        "instl": 1,
         "ext": {
           "bidder": {
             "app_store_id": "123",
@@ -35,17 +34,12 @@
     ],
     "user": {
       "buyeruid": "123"
-    },
-    "ext": {
-      "prebid": {
-        "debug": true
-      }
     }
   },
   "httpCalls": [
     {
       "expectedRequest": {
-        "uri": "https://liftoff.io/bit/t",
+        "uri": "https://vungle.io/bit/t",
         "headers": {
           "Content-Type": ["application/json"],
           "Accept": ["application/json"],
@@ -54,7 +48,8 @@
         "body": {
           "id": "test-request-id",
           "app": {
-            "id": "123"
+            "id": "123",
+            "bundle": "com.prebid"
           },
           "device": {
             "ifa": "87857b31-8942-4646-ae80-ab9c95bf3fab"
@@ -77,7 +72,6 @@
                   "foo": "bar"
                 }
               },
-              "instl": 1,
               "ext": {
                 "prebid": null,
                 "bidder": {
@@ -94,60 +88,16 @@
           ],
           "user": {
             "buyeruid": "123"
-          },
-          "ext": {
-            "prebid": {
-              "debug": true
-            }
           }
         },
         "impIDs":["test-imp-id"]
       },
       "mockResponse": {
-        "status": 200,
-        "body": {
-          "id": "test-request-id",
-          "seatbid": [
-            {
-              "seat": "seat-id",
-              "bid": [
-                {
-                  "id": "1",
-                  "impid": "test-imp-id",
-                  "adid": "11110126",
-                  "price": 0.500000,
-                  "adm": "some-test-ad",
-                  "crid": "test-crid",
-                  "h": 250,
-                  "w": 300
-                }
-              ]
-            }
-          ],
-          "cur": "USD"
-        }
+        "status": 204,
+        "body": {}
       }
     }
   ],
-  "expectedBidResponses": [
-    {
-      "currency": "USD",
-      "bids": [
-        {
-          "bid": {
-            "id": "1",
-            "impid": "test-imp-id",
-            "adid": "11110126",
-            "price": 0.5,
-            "adm": "some-test-ad",
-            "crid": "test-crid",
-            "w": 300,
-            "h": 250
-          },
-          "type": "video",
-          "seat": "seat-id"
-        }
-      ]
-    }
-  ]
+  "expectedBidResponses": [],
+  "expectedMakeBidsErrors": []
 }
