+++ adapters/vungle/vungletest/supplemental/response_code_400.json
@@ -1,8 +1,8 @@
 {
   "mockBidRequest": {
     "id": "test-request-id",
-    "site": {
-      "name": "com.prebid"
+    "app": {
+      "bundle": "com.prebid"
     },
     "device": {
       "ifa": "87857b31-8942-4646-ae80-ab9c95bf3fab"
@@ -25,9 +25,6 @@
           }
         },
         "ext": {
-          "prebid": {
-            "is_rewarded_inventory": 1
-          },
           "bidder": {
             "app_store_id": "123",
             "placement_reference_id": "123"
@@ -42,7 +39,7 @@
   "httpCalls": [
     {
       "expectedRequest": {
-        "uri": "https://liftoff.io/bit/t",
+        "uri": "https://vungle.io/bit/t",
         "headers": {
           "Content-Type": ["application/json"],
           "Accept": ["application/json"],
@@ -51,7 +48,8 @@
         "body": {
           "id": "test-request-id",
           "app": {
-            "id": "123"
+            "id": "123",
+            "bundle": "com.prebid"
           },
           "device": {
             "ifa": "87857b31-8942-4646-ae80-ab9c95bf3fab"
@@ -75,9 +73,7 @@
                 }
               },
               "ext": {
-                "prebid": {
-                  "is_rewarded_inventory": 1
-                },
+                "prebid": null,
                 "bidder": {
                   "app_store_id": "123",
                   "placement_reference_id": "123"
@@ -97,50 +93,16 @@
         "impIDs":["test-imp-id"]
       },
       "mockResponse": {
-        "status": 200,
-        "body": {
-          "id": "test-request-id",
-          "seatbid": [
-            {
-              "seat": "seat-id",
-              "bid": [
-                {
-                  "id": "1",
-                  "impid": "test-imp-id",
-                  "adid": "11110126",
-                  "price": 0.500000,
-                  "adm": "some-test-ad",
-                  "crid": "test-crid",
-                  "h": 250,
-                  "w": 300
-                }
-              ]
-            }
-          ],
-          "cur": "USD"
-        }
+        "status": 400,
+        "body": {}
       }
     }
   ],
-  "expectedBidResponses": [
+  "expectedBidResponses": [],
+  "expectedMakeBidsErrors": [
     {
-      "currency": "USD",
-      "bids": [
-        {
-          "bid": {
-            "id": "1",
-            "impid": "test-imp-id",
-            "adid": "11110126",
-            "price": 0.5,
-            "adm": "some-test-ad",
-            "crid": "test-crid",
-            "w": 300,
-            "h": 250
-          },
-          "type": "video",
-          "seat": "seat-id"
-        }
-      ]
+      "value": "Unexpected status code: 400.*",
+      "comparison": "regex"
     }
   ]
 }
