+++ adapters/vungle/vungletest/supplemental/vungle_ext_check.json
@@ -25,11 +25,9 @@
           }
         },
         "ext": {
-          "prebid": {
-            "is_rewarded_inventory": 1
-          },
           "bidder": {
-            "placement_reference_id": "78910"
+            "app_store_id": "123",
+            "placement_reference_id": "123"
           }
         }
       }
@@ -41,7 +39,7 @@
   "httpCalls": [
     {
       "expectedRequest": {
-        "uri": "https://liftoff.io/bit/t",
+        "uri": "https://vungle.io/bit/t",
         "headers": {
           "Content-Type": ["application/json"],
           "Accept": ["application/json"],
@@ -50,6 +48,7 @@
         "body": {
           "id": "test-request-id",
           "app": {
+            "id": "123",
             "bundle": "com.prebid"
           },
           "device": {
@@ -58,7 +57,7 @@
           "imp": [
             {
               "id": "test-imp-id",
-              "tagid": "78910",
+              "tagid": "123",
               "video": {
                 "mimes": [
                   "video/mp4"
@@ -74,16 +73,15 @@
                 }
               },
               "ext": {
-                "prebid": {
-                  "is_rewarded_inventory": 1
-                },
+                "prebid": null,
                 "bidder": {
-                  "placement_reference_id": "78910"
+                  "app_store_id": "123",
+                  "placement_reference_id": "123"
                 },
                 "vungle": {
-                  "app_store_id": "",
                   "bid_token": "123",
-                  "placement_reference_id": "78910"
+                  "app_store_id": "123",
+                  "placement_reference_id": "123"
                 }
               }
             }
@@ -93,14 +91,52 @@
           }
         },
         "impIDs":["test-imp-id"]
+      },
+      "mockResponse": {
+        "status": 200,
+        "body": {
+          "id": "test-request-id",
+          "seatbid": [
+            {
+              "seat": "seat-id",
+              "bid": [
+                {
+                  "id": "1",
+                  "impid": "test-imp-id",
+                  "adid": "11110126",
+                  "price": 0.500000,
+                  "adm": "some-test-ad",
+                  "crid": "test-crid",
+                  "h": 250,
+                  "w": 300
+                }
+              ]
+            }
+          ],
+          "cur": "USD"
+        }
       }
     }
   ],
-  "expectedBidResponses": [],
-  "expectedMakeBidsErrors": [
+  "expectedBidResponses": [
     {
-      "value": "Unexpected status code: 0*",
-      "comparison": "regex"
+      "currency":"USD",
+      "bids":[
+        {
+          "bid": {
+            "id": "1",
+            "impid": "test-imp-id",
+            "adid": "11110126",
+            "price": 0.500000,
+            "adm": "some-test-ad",
+            "crid": "test-crid",
+            "h": 250,
+            "w": 300
+          },
+          "type": "video",
+          "seat": "seat-id"
+        }
+      ]
     }
   ]
 }
