+++ openrtb_ext/imp_liftoff.go
@@ -1,6 +1,6 @@
 package openrtb_ext
 
-type ImpExtLiftoff struct {
+type ImpExtVungle struct {
 	BidToken       string `json:"bid_token"`
 	PubAppStoreID  string `json:"app_store_id"`
 	PlacementRefID string `json:"placement_reference_id"`
