+++ static/bidder-info/vungle.yaml
